### 2022-05-11 #4386 @editorialbot
[PRE REVIEW]: TimeseriesSurrogates.jl: a Julia package for generating surrogate data
**Submitting author:** @kahaaga (<a href="http://orcid.org/0000-0001-6880-8725">Kristian Agasøster Haaga</a>)
**Repository:** https://github.com/JuliaDynamics/TimeseriesSurrogates.jl
**Branch with paper.md** (empty if default branch): JOSSpaper
**Version:** v2.0.0
