### 2022-05-21 adf0d2d
Add DOIs, and fix theiler citation
Proper indentation
