+++ paper/paper.md
@@ -27,22 +27,22 @@ bibliography: paper.bib
 ---
 
 # Introduction
-The method of surrogate data is a way to generate data that preserve one or more statistical or dynamical properties of a given timeseries, but are otherwise randomized. Surrogate time series methods have widespread use in null hypothesis testing in nonlinear dynamics, for null hypothesis testing in causal inference, or for the more general case of producing synthetic data with similar statistical properties as an original signal. Originally introduced by @Theiler:1991 to test for nonlinearity in time series, numerous surrogate methods aimed preserving different properties of the original signal have since emerged; for a review, see @Lancaster:2018.
+The method of surrogate data is a way to generate data that preserve one or more statistical or dynamical properties of a given timeseries, but are otherwise randomized. Surrogate time series methods have widespread use in null hypothesis testing in nonlinear dynamics, for null hypothesis testing in causal inference, or for the more general case of producing synthetic data with similar statistical properties as an original signal. Originally introduced by @Theiler:1992 to test for nonlinearity in time series, numerous surrogate methods aimed preserving different properties of the original signal have since emerged; for a review, see @Lancaster:2018.
 
 A simple example of an application of surrogates would be to distinguish whether a given timeseries `x` can be represented via a linear noise process, or not. The latter case can be an indication that the timeseries may represent deterministic nonlinear dynamics with additional noise. A simple way to test for this hypothesis would be to generate new timeseries from `x` that conserve the power spectrum of `x` (which is a defining feature of linear stochastic processes). Then, a discriminatory statistic, such as the correlation dimension or the auto-mutual-information [@Lancaster:2018] is computed for `x`, but also for thousands of surrogates from `x`. The discriminatory statistic of the surrogates provides a distribution of possible values, and if the value for `x` is well within the distribution spread, then `x` satisfies the null hypothesis (here, that `x` can be approximated as a linear stochastic process).
 
 # Statement of need
-Surrogate data has been used in several thousand publications so far (the citation number of @Theiler:1991 is more than 4,000) and hence the community is in clear need of such methods. Existing software packages for surrogate generation provide much fewer methods than available in the literature, with less-than optimal performance (see Comparison section below), and without allowing reproducible generation of surrogates. TimeseriesSurrogates.jl provides more than double the amount of methods given by other packages, with runtimes similar to and up to an order of magnitude faster than existing surrogate packages in other languages. Equally importantly, TimeseriesSurrogates.jl provides a framework that is tested via continuous integration, and is easy to extend via open source contributions.
+Surrogate data has been used in several thousand publications so far (the citation number of @Theiler:1992 is more than 4,000) and hence the community is in clear need of such methods. Existing software packages for surrogate generation provide much fewer methods than available in the literature, with less-than optimal performance (see Comparison section below), and without allowing reproducible generation of surrogates. TimeseriesSurrogates.jl provides more than double the amount of methods given by other packages, with runtimes similar to and up to an order of magnitude faster than existing surrogate packages in other languages. Equally importantly, TimeseriesSurrogates.jl provides a framework that is tested via continuous integration, and is easy to extend via open source contributions.
 
 # Available surrogate methods
 
 | Method | Description | Reference |
 |---|---|---|
 | `AutoRegressive`  | Autoregressive model based surrogates. | |
-| `RandomShuffling` | Random shuffling of individual data points. | @Theiler:1991 |
-| `BlockShuffle`  | Random shuffling of blocks of data points.  | @Theiler:1991 |
[…]
