+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={748},
   pages={1--60},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.physrep.2018.06.001},
 }
 
 @article{Bezanson:2017,
@@ -16,14 +17,22 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/141000671},
 }
 
-@techreport{Theiler:1991,
-  title={Testing for nonlinearity in time series: the method of surrogate data},
-  author={Theiler, James and Galdrikian, B and Longtin, Andr  and Eubank, S and Farmer, J Doyne},
-  year={1991},
-  institution={Los Alamos National Lab., NM (United States)}
+@article{Theiler:1992,
+  url = {https://doi.org/10.1016%2F0167-2789%2892%2990102-s},
+  year = 1992,
+  month = {sep},
+  publisher = {Elsevier {BV}},
+  volume = {58},
+  number = {1-4},
+  pages = {77--94},
+  author = {James Theiler and Stephen Eubank and Andr{\'{e}} Longtin and Bryan Galdrikian and J. Doyne Farmer},
+  title = {Testing for nonlinearity in time series: the method of surrogate data},
+  journal = {Physica D: Nonlinear Phenomena},
+  doi = {10.1016/0167-2789(92)90102-s},
 }
 
 @article{Theiler:1994,
@@ -34,7 +43,8 @@
   number={1-2},
   pages={335--341},
   year={1994},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/0375-9601(94)91096-0},
 }
     
 @article{SchreiberSchmitz:1996,
@@ -45,7 +55,8 @@
   number={4},
   pages={635},
   year={1996},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevlett.77.635},
 }
 
 @article{Miralles2015,
@@ -56,7 +67,8 @@
   number={3},
   pages={1595--1603},
   year={2015},
-  publisher={Acoustical Society of America}
+  publisher={Acoustical Society of America},
+  doi = {10.1121/1.4929694},
 }
         
 @article{Nakamura:2006,
@@ -67,7 +79,8 @@
   number={2},
   pages={026205},
   year={2006},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physreve.74.026205},
 }
         
 @article{Keylock:2006,
@@ -78,7 +91,8 @@
   number={3},
   pages={036707},
   year={2006},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physreve.73.036707},
 }
 
 @article{Small:2001,
@@ -89,7 +103,8 @@
   number={18},
   pages={188101},
   year={2001},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevlett.87.188101},
 }
 
 @article{Schmitz:1999,
@@ -100,7 +115,8 @@
   number={4},
   pages={4044},
   year={1999},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physreve.59.4044},
 }
 @article{Palus:2008,
   title={Bootstrapping multifractals: Surrogate data from random cascades on wavelet dyadic trees},
@@ -110,7 +126,8 @@
   number={13},
   pages={134101},
   year={2008},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevlett.101.134101},
 }
 
 @article{Ortega:1998,
@@ -121,7 +138,8 @@
   number={20},
   pages={4345},
[…]
