### 2022-05-21 #4386 @lucaferranti
I could review this (I have another JOSS review ongoing atm, but it's almost concluded, so it shouldn't be a problem)
