### 2022-05-22 #4386 @dawbarton
I'd be interested in reviewing this as it has relevance for my own work. It'll probably be 3-4 weeks before I can do the review though.
