### 2022-05-25 #4414 @lucaferranti
@jarvist I didn't know reviewers have to generate their checklist, is it new? Anyway I generated my checklist and looks like I can edit, so I guess everything is setup? :)
