### 2022-06-08 #4414 @lucaferranti
Hi @kahaaga (cc @jarvist ) overall a good submission! Here is the first batch of comments related to the code

**Note**: This PR is based on the code and docs in the main branch. I noticed you have a PR "JOSS paper" that with a quick look seems to address some of the comments I have below, could you clarify the status of that PR? Maybe it can be merged if it's ready?

### Functionality

- [ ] At the moment installation instructions are missing, it would be good to have those both in the README and in the docs homepage (see below)
- [ ] The paper has some benchmarks, it would be good to include the code to reproduce those benchmarks. I guess there are several way to approach this, one idea could be to have a benchmarks article also in the documentation, where code snippets and results are included and explained and/or having a benchmark folder with e.g. 1) a julia file for the julia experiments 2) a matlab file for the matlab experiments 3) a postprocessing file for generating the results and graphs (this could be together with 1).

### Documentation
- [ ] Several requirements are currently, missing e.g. installation instructions, contributing guidelines etc., please see the checkpoints in the list above.
- [ ] Regarding the structure, currently the homepage goes straight into the API docs, this could be a little overwhelming for a newcomer, I would suggest the homepage of the docs to include the followings (but feel free to add remove items if you prefer), it would also be good to have a more or less similar structure in the readme too.
  1. A general overview of the package
  2. installation instructions
  3. A small quickstart example, something small and easy to get the taste of the package
- [ ] Currently the readme has only the link to the docs of the dev version but not to the latest stable release.

### Other
a few minor comments

- [ ] the CI badge in the readme points to travis, but looking at the PRs you seem to use github actions (indeed the latest action on travis seems to be one year old). I guess the travis badge is a leftover? In that case the `.travis.yml` could be removed too probably?
- [ ] (very minor, feel free to ignore if you like) I noticed you are already using codecov to compute the coverage. Since you have it already setup, have you considered adding a badge and link to the coverage report in the readme? The coverage is 88%, so it would also look good in the repo :)

This was a first batch of comments, I'll next read the paper more carefully and check I can reproduce the examples in the docs.
