### 2022-06-08 eab14e3
JOSS paper (#62)
* Add draft of JOSS paper

* Contribution guide and installation

* Acknowledgements

* Cosmetics

* Fixed to contribution guide

* Add note on docstring in contribution guide

* add GD funding

* add design discussion

* Move installation to overview page

* MATLAB methods (DELETE THIS AT THE END)

This file contains the MATLAB methods that the Lancaster group has implemented, and I add it here simply as a table of things they have, so that we can compare easily in our paper. If we can run MATLAB code, this can also be used as a benchmark.

* Update Project.toml

* No need fo
[…]
