+++ paper/paper.md
@@ -1,0 +1,119 @@
+---
+title: 'TimeseriesSurrogates.jl: a Julia package for generating surrogate data'
+tags:
+  - Julia
+  - surrogate data
+  - time series
+  - nonlinear time series analysis
+  - hypothesis testing
+authors:
+  - name: Kristian Agasøster Haaga
+    orcid: 0000-0001-6880-8725
+    affiliation: "1, 2, 3"
+  - name: George Datseris
+    orcid: 0000-0002-6427-2385
+    affiliation: "4"
+affiliations:
+  - name: Department of Earth Science, University of Bergen, Bergen, Norway
+    index: 1
+  - name: K. G. Jebsen Centre for Deep Sea Research, Bergen, Norway
+    index: 2
+  - name: Bjerknes Centre for Climate Research, Bergen, Norway
+    index: 3
+  - name:  Max Planck Institute for Meteorology, Hamburg, Germany
+    index: 4
+date: 9 March 2020
+bibliography: paper.bib
+---
+
+# Introduction
+The method of surrogate data is a way to generate data that preserve one or more statistical or dynamical properties of a given timeseries, but are otherwise randomized. Surrogate time series methods have widespread use in null hypothesis testing in nonlinear dynamics, for null hypothesis testing in causal inference, or for the more general case of producing synthetic data with similar statistical properties as an original signal. Originally introduced by @Theiler:1992 to test for nonlinearity in time series, numerous surrogate methods aimed preserving different properties of the original signal have since emerged; for a review, see @Lancaster:2018.
+
+A simple example of an application of surrogates would be to distinguish whether a given timeseries `x` can be represented via a linear noise process, or not. The latter case can be an indication that the timeseries may represent deterministic nonlinear dynamics with additional noise. A simple way to test for this hypothesis would be to generate new timeseries from `x` that conserve the power spectrum of `x` (which is a defining feature of linear stochastic processes). Then, a discriminatory statistic, such as the correlation dimension or the auto-mutual-information [@Lancaster:2018] is computed for `x`, but also for thousands of surrogates from `x`. The discriminatory statistic of the surrogates provides a distribution of possible values, and if the value for `x` is well within the distribution spread, then `x` satisfies the null hypothesis (here, that `x` can be approximated as a linear stochastic process).
+
+# Statement of need
+Surrogate data has been used in several thousand publications so far (the citation number of @Theiler:1992 is more than 4,000) and hence the community is in clear need of such methods. Existing software packages for surrogate generation provide much fewer methods than available in the literature, with less-than optimal performance (see Comparison section below), and without allowing reproducible generation of surrogates. TimeseriesSurrogates.jl provides more than double the amount of methods given by other packages, with runtimes similar to and up to an order of magnitude faster than existing surrogate packages in other languages. Equally importantly, TimeseriesSurrogates.jl provides a framework that is tested via continuous integration, and is easy to extend via open source contributions.
+
+# Available surrogate methods
+
+| Method | Description | Reference |
+|---|---|---|
+| `AutoRegressive`  | Autoregressive model based surrogates. | |
+| `RandomShuffling` | Random shuffling of individual data points. | @Theiler:1992 |
+| `BlockShuffle`  | Random shuffling of blocks of data points.  | @Theiler:1992 |
+| `CircShift`  | Circularly shift the signal.  | |
+| `RandomFourier`  | Randomization of phases of Fourier transform of the signal.  | @Theiler:1992 |
[…]
