+++ paper/paper.bib
@@ -1,0 +1,155 @@
+@article{Lancaster:2018,
+  title={Surrogate data for hypothesis testing of physical systems},
+  author={Lancaster, Gemma and Iatsenko, Dmytro and Pidde, Aleksandra and Ticcinelli, Valentina and Stefanovska, Aneta},
+  journal={Physics Reports},
+  volume={748},
+  pages={1--60},
+  year={2018},
+  publisher={Elsevier},
+  doi={10.1016/j.physrep.2018.06.001},
+}
+
+@article{Bezanson:2017,
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal={SIAM review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM},
+  doi={10.1137/141000671},
+}
+
+@article{Theiler:1992,
+  url = {https://doi.org/10.1016%2F0167-2789%2892%2990102-s},
+  year = 1992,
+  month = {sep},
+  publisher = {Elsevier {BV}},
+  volume = {58},
+  number = {1-4},
+  pages = {77--94},
+  author = {James Theiler and Stephen Eubank and Andr{\'{e}} Longtin and Bryan Galdrikian and J. Doyne Farmer},
+  title = {Testing for nonlinearity in time series: the method of surrogate data},
+  journal = {Physica D: Nonlinear Phenomena},
+  doi = {10.1016/0167-2789(92)90102-s},
+}
+
+@article{Theiler:1994,
+  title={On the evidence for low-dimensional chaos in an epileptic electroencephalogram},
+  author={Theiler, James},
+  journal={Physics Letters A},
+  volume={196},
+  number={1-2},
+  pages={335--341},
+  year={1994},
+  publisher={Elsevier},
+  doi={10.1016/0375-9601(94)91096-0},
+}
+    
+@article{SchreiberSchmitz:1996,
+  title={Improved surrogate data for nonlinearity tests},
+  author={Schreiber, Thomas and Schmitz, Andreas},
+  journal={Physical review letters},
+  volume={77},
+  number={4},
+  pages={635},
+  year={1996},
+  publisher={APS},
+  doi={10.1103/physrevlett.77.635},
+}
+
+@article{Miralles2015,
+  title={Characterization of the complexity in short oscillating time series: An application to seismic airgun detonations},
+  author={Miralles, R and Carrion, A and Looney, D and Lara, G and Mandic, D},
+  journal={The Journal of the Acoustical Society of America},
+  volume={138},
+  number={3},
+  pages={1595--1603},
+  year={2015},
+  publisher={Acoustical Society of America},
+  doi = {10.1121/1.4929694},
+}
+        
+@article{Nakamura:2006,
+  title={Testing for nonlinearity in irregular fluctuations with long-term trends},
+  author={Nakamura, Tomomichi and Small, Michael and Hirata, Yoshito},
+  journal={Physical Review E},
+  volume={74},
+  number={2},
+  pages={026205},
+  year={2006},
+  publisher={APS},
+  doi={10.1103/physreve.74.026205},
+}
+        
+@article{Keylock:2006,
+  title={Constrained surrogate time series with preservation of the mean and variance structure},
+  author={Keylock, CJ},
+  journal={Physical Review E},
+  volume={73},
+  number={3},
+  pages={036707},
+  year={2006},
+  publisher={APS},
+  doi={10.1103/physreve.73.036707},
+}
+
+@article{Small:2001,
+  title={Surrogate test for pseudoperiodic time series data},
[…]
