+++ docs/src/index.md
@@ -69,4 +69,11 @@ fig = surroplot(ts, s)
 save("surroplot.png", fig); # hide
 ```
 
-## References
+## Installation
+
+TimeseriesSurrogates is a registered Julia package. To install the latest version, run the following in your Julia console.
+
+```julia
+import Pkg; Pkg.add("TimeseriesSurrogates")
+```
+
