+++ README.md
@@ -1,6 +1,6 @@
 # TimeseriesSurrogates.jl
 
-[![Build Status](https://travis-ci.org/JuliaDynamics/TimeseriesSurrogates.jl.svg?branch=master)](https://travis-ci.org/JuliaDynamics/TimeseriesSurrogates.jl)
+[![CI](https://github.com/juliadynamics/TimeseriesSurrogates.jl/workflows/CI/badge.svg)](https://github.com/JuliaDynamics/TimeseriesSurrogates.jl/actions)
 [![codecov](https://codecov.io/gh/JuliaDynamics/TimeseriesSurrogates.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/JuliaDynamics/TimeseriesSurrogates.jl)
 [![](https://img.shields.io/badge/docs-online-blue.svg)](https://JuliaDynamics.github.io/TimeseriesSurrogates.jl/dev)
 
