### 2022-06-08 #4414 @kahaaga
Hey @lucaferranti! Thanks for the initial comments.

> This PR is based on the code and docs in the main branch. I noticed you have a PR "JOSS paper" that with a quick look seems to address some of the comments I have below, could you clarify the status of that PR? Maybe it can be merged if it's ready?

When submitting the paper, I had the option to select a branch to be considered for the publication. I chose the `JOSSPaper` branch, so that's the latest version of the paper/docs/etc. However, I see that it's more convenient if everything is on the master branch, so I merged it now. The documentation badge in the repo readme now points to the updated documentation.

I think that should address most of your initial comments. In particular:

> The paper has some benchmarks, it would be good to include the code to reproduce those benchmarks. I guess there are several way to approach this, one idea could be to have a benchmarks article also in the documentation, where code snippets and results are included and explained and/or having a benchmark folder with e.g. 1) a julia file for the julia experiments 2) a matlab file for the matlab experiments 3) a postprocessing file for generating the results and graphs (this could be together with 1).

The benchmarks are now in the `paper/timing/code` folder of the repo.

> Regarding the structure, currently the homepage goes straight into the API docs, this could be a little overwhelming for a newcomer, I would suggest the homepage of the docs to include the followings (but feel free to add remove items if you prefer), it would also be good to have a more or less similar structure in the readme too. 1) A general overview of the package 2) installation instructions 3) A small quickstart example, something small and easy to get the taste of the package

Please have a look at the updated documentation. It now includes installation instructions, a brief overview of the package (with a link to the dedicated "what is a surrogate" page).

As for the quickstart example, there are already plentyful of examples in the documentation - one for most of the methods. We'd prefer to keep the documentation as minimal as possible, without unnecessary repetition. Maybe it'd be enough to include a sentence in the opening paragraphs of the docs that examples are found under the individual method pages?

>  the CI badge in the readme points to travis

The CI badge now points to Github Actions, which are currently used.

> (very minor, feel free to ignore if you like) I noticed you are already using codecov to compute the coverage. Since you have it already setup, have you considered adding a badge and link to the coverage report in the readme? The coverage is 88%, so it would also look good in the repo :)

A codecov badge is now included in the readme for the repo!
