+++ Project.toml
@@ -25,7 +25,7 @@ DSP = "^0.5.2, 0.6, ^1, 0.7"
 DelayEmbeddings = "2"
 Distances = "0.10, 1"
 Distributions = "0.21, 1, 0.23, 0.24, 0.25"
-Interpolations = "^0.12, ^1, 0.13"
+Interpolations = "^0.12, ^1, 0.13, 0.14"
 LombScargle = "1"
 Requires = "^0.5.2, ^1"
 StatsBase = "^0.32.0, ^1, 0.33"
