+++ paper/timing/code/timings.jl
@@ -1,38 +1,33 @@
-using Pkg; 
-Pkg.activate("./")
-Pkg.instantiate()
-using TimeseriesSurrogates, BenchmarkTools, Plots, Distributions, DynamicalSystems, Random, StatsPlots;
+# Activate the current directory and load packages.
+using Pkg;
+Pkg.activate(@__DIR__)
+using DelimitedFiles, TimeseriesSurrogates, BenchmarkTools, Plots, Distributions, 
+    DynamicalSystems, Random, StatsPlots;
+
+# Use GR backend for plotting.
 gr()
 
-# Here, we compute the mean time to generate a single surrogate for each of the methods 
-# implemented in TimeseriesSurrogates.jl at the time of submission to JOSS. 
-# We'll use the same test process as in Lancaster et al, and generate a 2000-pt long
-# time series with additive noise from a normal distribution with zero mean and 
-# standard deviation 0.3.
-f(t) = sin(2π*(t + 0.5*sin(2π*t/10))/3)
-npts = 2000
+x = readdlm("$(@__DIR__)/timeseries.csv")[:, 1]
 
-# Add some noise to the signal.
-rng = Random.MersenneTwister(1234)
-t = collect(1:npts) .+ rand(rng, npts)
-x = f.(t) .+ rand(rng, Normal(0, 0.3), npts)
-
-# Write to file, to import in matlab
-using DelimitedFiles
-writedlm("./paper/timing/code/data.csv",  x, ',')
+# Reproducibility (note: this is not possible using Lancaster et al.'s MATLAB code.)
+rng = Random.MersenneTwister(1234);
 
 # Pre-initialize all different surrogate types
+# ------------------------------------------
+
 # Shuffle based surrogates
 rs = surrogenerator(x, RandomShuffle(), rng)
 rp = surrogenerator(x, RandomFourier(true), rng)
 aaft = surrogenerator(x, AAFT(), rng)
 iaaft = surrogenerator(x, IAAFT(), rng)
 
-# Pseudoperiodic surrogates (PPS) 
-τ = estimate_delay(x, "ac_min")
-d = findmin(delay_fnn(x, τ))[2]
-ρ = noiseradius(x, d, τ, 0.025:0.025:0.5)
-pps = surrogenerator(x, PseudoPeriodic(d, τ, ρ, true))
+# Pseudoperiodic surrogates (PPS). Here, as in "timings.m", we use a subset of 500 time 
+# series points.
+x_pps = x[1:1000]
+τ = estimate_delay(x_pps, "ac_min")
+d = findmin(delay_fnn(x_pps, τ))[2]
+ρ = noiseradius(x_pps, d, τ, 0.025:0.025:0.5)
+pps = surrogenerator(x_pps, PseudoPeriodic(d, τ, ρ, true))
 
 # Collect surrogate generators and run once to trigger precompilation
 surrogens = [rs, rp, aaft, iaaft, pps];
@@ -41,7 +36,6 @@ surrogates = [s() for s in surrogens];
 # Define labels for plot.
 methods = ["RandomShuffle()", "RandomFourier(true)", "AAFT()", "IAAFT()", "PseudoPeriodic()"];
 n_methods = length(methods)
-
 
 function generate_surrs(surrogenerator, n)
     for i = 1:n
@@ -65,12 +59,14 @@ function time_n(methods; n = 100)
 end
 time_n(surrogens, n = 1);
 
-# Compute average time to generate surrogate over 30 realizations of each surrogate type
-timings = time_n(surrogens, n = 100); 
+# Like in `timings.m`, compute average time to generate surrogate over 500 
+# realizations of each surrogate type
[…]
