### 2022-08-09 bbfb490
Add instructions for running timing code
