### 2022-08-09 40bbe19
Remove unnecessary dependencies
