### 2022-08-09 eea1716
Adjust plot limits and ticks to better highlight differences
