### 2022-08-09 4daa648
Generate `nsurr` surrogates, and normalize timing to `nsurr`
