### 2022-08-09 5b2ce7f
Generate time series in separate script
