+++ paper/timing/code/Project.toml
@@ -5,3 +5,6 @@ DynamicalSystems = "61744808-ddfa-5f27-97ff-6e42cc95d634"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
 TimeseriesSurrogates = "c804724b-8c18-5caa-8579-6025a0767c70"
+
+[compat]
+TimeseriesSurrogates = "2"
