+++ paper/paper.md
@@ -107,7 +107,7 @@ The average time to generate surrogates in TimeseriesSurrogates.jl is in the bes
 Timings for commonly used surrogate methods that are common to both libraries are shown in Figure 1.
 Additionally, because TimeseriesSurrogates.jl provides many more methods not implemented in other packages, a comprehensive comparison of runtimes is not possible, but due to our optimized surrogate generators, we expect good performance relative to future implementations in other languages.
 
-![Figure 1: Mean time (in seconds, based on 100 realizations) to generate a random permutation (rp), Fourier transform (ft), amplitude-adjusted Fourier transform (aaft), iterated aaft (iaaft) and pseudoperiodic (pps) surrogate using a pre-initialized generators with default parameters, and using a maximum of 100 iterations for the IAAFT algorithm. MATLAB timings are generated using the code provided by @Lancaster:2018. Note: timings for the pseudoperiodic surrogates in MATLAB include embedding lag and dimension finding, which has been included in the preprocessing step in the Julia version. Scripts to reproduce Julia and MATLAB timings are available in the GitHub repo for this paper. ](figs/timings.png)
+![Figure 1: Mean time (in seconds, based on 500 surrogate time series realizations) to generate a 2000-point random permutation (rp), Fourier transform (ft), amplitude-adjusted Fourier transform (aaft), iterated aaft (iaaft) and pseudoperiodic (pps) surrogate using a pre-initialized generators with default parameters, and using a maximum of 100 iterations for the IAAFT algorithm. MATLAB timings are generated using the code provided by @Lancaster:2018. Note: timings for the pseudoperiodic surrogates in MATLAB include embedding lag and dimension finding, which has been included in the preprocessing step in the Julia version. Scripts to reproduce Julia and MATLAB timings are available in the GitHub repo for this paper. ](figs/timings.png)
 
 # Acknowledgements
 
