+++ paper/timing/code/timings.m
@@ -1,27 +1,30 @@
-data = readtable("data.csv");
-x = data{:, 1};
-n = 100;
+run generate_timeseries.m
+
+nsurr = 500;
+
+% Time methods
+tic
+surrogate(x, nsurr, 'RP', 0, 1);
+t_rp = toc / nsurr;
 
 tic
-surrogate(x, n, 'RP', 0, 1);
-t_rp = toc / n;
+surrogate(x, nsurr, 'FT', 0, 1);
+t_ft = toc / nsurr;
 
 tic
-surrogate(x, n, 'FT', 0, 1);
-t_ft = toc / n;
+surrogate(x, nsurr, 'AAFT', 0, 1);
+t_aaft = toc / nsurr;
 
 tic
-surrogate(x, n, 'AAFT', 0, 1);
-t_aaft = toc / n;
+surrogate(x, nsurr, 'IAAFT2', 0, 1);
+t_iaaft = toc / nsurr;
 
 tic
-surrogate(x, n, 'IAAFT2', 0, 1);
-t_iaaft = toc / n;
+% Use a subset of the time series for the PPS method (it is slow)
+x_pps = x(1:1000);
+surrogate(x_pps, nsurr, 'PPS', 0, 1);
+t_pps = toc / nsurr;
 
-tic
-surrogate(x, n, 'PPS', 0, 1);
-t_pps = toc / n;
-
-
-[t_rp, t_ft, t_aaft, t_iaaft, t_pps];
-
+% Save MATLAB timings for import in Julia script.
+matlab_times = [t_rp, t_ft, t_aaft, t_iaaft, t_pps];
+writematrix(matlab_times.', './matlab_timings.csv')
