### 2022-08-09 #4414 @kahaaga
@jarvist Sorry for the delay!

We have now addressed @dawbarton's comments, which have been merged to the `main` branch.

> I tried the timing code - it was a bit annoying as it can't just be run without modification. Personally, I would separate out the data generation or the plotting code (or both) because at the moment you have to copy-and-paste the code in piecemeal because there is a circular dependency. (Matlab needs to use the same data that is generated in the Julia script but the Julia script also plots pre-saved Matlab results that can only be regenerated once you've got the data.) Also, the hardcoded paths are a bit annoying - you could use @__FILE__ or @__DIR__ instead.

Good point. I now use `@__DIR__` for file paths, and reorganised the timing scripts so that tasks are delegated a bit differently between MATLAB and Julia. Now, first run the MATLAB script (which will generate example data), then the Julia script (which will do the comparison and generate the plot for the paper).

I also increased the number of surrogates generated for the comparison for a bit more reliable timings (@jarvist note: the paper pdf must be re-generated to include the update timing plot).

> It might be worth putting a little more disambiguation of the term "surrogates" in the package readme / docs. It's a bit of an overloaded term with surrogate models being a significant area of interest at the moment (but completely different).

The wording in the documentation was slightly adjusted to highlight that we are dealing with *time series* surrogates.
