### 2022-08-18 #4414 @lucaferranti
@kahaaga @jarvist there is an error when generating the pdf, not sure if it's a bot issue or paper issue, can you have a look at it?
