### 2022-08-18 #4414 @kahaaga
@lucaferranti I originally submitted the paper referring to the `JOSSPaper` branch, which no longer exists, because it was merged to `main` and deleted (see discussion above).

The [error](https://github.com/openjournals/joss-papers/actions/runs/2880454779?check_suite_focus=true) seems to be related to the fact that this branch no longer exists.

Maybe running `@editorialbot set main as branch`, then re-generating the pdf will work?
