### 2022-08-18 #4414 @lucaferranti
meanwhile @kahaaga, overview software and docs look good to me. As a small nitpick, currently the installation instructions in the docs homepage are last. This seems a little strange to me, I would suggest to move that section before the `API` section.

I will check the paper once I can generate the pdf.
