### 2022-08-18 #4414 @lucaferranti
> Maybe running `@editorialbot set main as branch`, then re-generating the pdf will work?

I don't think a reviewer should be allowed to change the branch where the paper is generated from, that sounds like something only author and/or editor should be allowed to do, can you try it?
