### 2022-08-18 238ce1d
Move installation instructions to start (#129)
* Move installation instructions to start

* Wording
