### 2022-08-18 #4414 @lucaferranti
@kahaaga I've now gone through the paper and it seems good to me. As a minor comment, I think the full list of papers needs not be included in the paper, as it's more an API documentation (which is already well covered in the docs). That being said, it doesn't hurt either to be there if you prefer to keep it.

@jarvist this concludes my review, I recommend the paper for publication.
