### 2022-08-22 #4414 @jarvist
I'm happy with the corrected version of the paper & documentation, and given your response to the reviewers comments, I can heartily recommend acceptance.

An editor in chief should be along to take up the process from here.

Thankyou @dawbarton and @lucaferranti for your efforts!
