### 2022-08-22 #4414 @jarvist
@Datseris Oooh, that's not good! If you are happy with your version v2.0.0, you should be able to push that to Zenodo or similar and generate the DOI.
