### 2022-08-22 #4414 @Datseris
@jarvist I'm on it. I'll publish a version 2.0.1 that includes the documentation improvements that the revierwers suggested. Then i'll also store this to Zenodo and provide a new DOI here!
