### 2022-08-22 #4414 @Datseris
@jarvist here is the doi: 10.5281/zenodo.7014633  , associated with version `v2.0.1` of the software.
