### 2022-09-10 #4414 @Datseris
@arfon I am submitting a proposal to the EU this Tuesday, and I guess it would be beneficial for me to include this paper in my publication lists since it is accepted. Given that the bot has some difficulties, do you know perhaps what will be the assigned DOI for this paper so I can use that in my proposal? Thanks!
