### 2022-09-12 #4414 @arfon
@xuanxu – can you take a look at the error from this submission? It looks like it might be associated with the JATS file uploads? https://github.com/openjournals/joss-papers/runs/8283823992?check_suite_focus=true#logs
