### 2022-09-12 #4414 @jarvist
I wonder if it's the `Andr{\'{e}}` markup in the author list, if Theiler:1992 is the one it's triggering on. But then Lucio:2012 should also be causing problems.
