### 2022-09-12 #4414 @xuanxu
I think the error is caused by a [reference](https://github.com/JuliaDynamics/TimeseriesSurrogates.jl/blob/main/paper/paper.md?plain=1#L50) to `Theiler:1991` in the paper that can't be found in the bib file
