### 2022-09-12 #4414 @kahaaga
@jarvist The year was indeed wrong for one of the references to the Theiler paper. It it fixed in the `main` branch on the repo now.
